requirements.txt
# Core ML
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
accelerate>=0.25.0
peft>=0.7.0
trl>=0.7.0
bitsandbytes>=0.41.0

# Vision
ultralytics>=8.0.0
Pillow>=10.0.0
opencv-python>=4.8.0

# Data
pycocotools>=2.0.7
datasets>=2.14.0

# Evaluation & Visualization
scikit-learn>=1.3.0
nltk>=3.8.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
wandb>=0.15.0
jsonlines>=4.0.0
